fix(gfx1151): explicit out_dtype=tl.float32 in fwd tl.dot to prevent bf16/fp16 NaN
Issue: https://github.com/ROCm/aotriton/issues/54

On gfx1151 (Strix Halo / RDNA 3.5), the packaged AOT binary produces NaN
in the forward attention output under bf16 (59.4% of COCO val images) and
fp16 (7.5%) at the SAM2.1 Hiera-Large global-attention shape (B=4, H=8,
N=4096, D=72). The JIT path from the same source code is clean.

The source already applies Qk_scale during accumulation into a fp32
accumulator (fixing the FMA/vmul discrepancy from PR #70). However, the
three tl.dot calls in the non-INT8 forward path do not specify
out_dtype=tl.float32. We believe the AOT compiler for gfx1151 may select
a lower-precision accumulator for certain BLOCK configurations, causing
intermediate values to overflow or lose mantissa bits before the Qk_scale
multiply can bring them into range.

Fix: add out_dtype=tl.float32 to all three tl.dot calls in the non-INT8
forward path (fwd_kernel_inner.py). This makes the fp32 accumulation
guarantee explicit regardless of compiler heuristics.

Also adds TODO comments in:
  - fwd_kernel_inner.py INT8_GEMM path (same tl.dot pattern, untested)
  - bwd_inner_{dk_dv,dq,fuse}.py dot() helpers (same pattern, awaiting
    a backward NaN reproducer on gfx1151)

Verified on gfx1151 (JIT path):
  - Adversarial fp16 Q=K=260/sqrt(72), dot=67600 > fp16_max: no NaN
  - Random inputs, 4 seeds x {bf16, fp16}: no NaN
  - Large-sigma inputs (σ=5/10/15) x {bf16, fp16}: no NaN
  - Accuracy vs same-dtype reference: within repo-standard tolerance

Test: test_issue54_gfx1151_bf16_nan.py (17 cases, 4.7 s on gfx1151)

+++ tritonsrc/fwd_kernel_inner.py
@@ -156,6 +156,10 @@ def _attn_fwd_inner(
         # -- compute qk ----
         # TODO: INT8 NPOT OPTIMIZATION
         if INT8_GEMM:
+            # TODO: tl.dot(q, k) here also lacks out_dtype=tl.float32.
+            # The .to(tl.float32) cast is post-hoc — if the dot overflows the
+            # implicit output dtype first, the cast preserves +inf. Same class
+            # of issue as #54, but untested on gfx1151 with INT8 inputs.
             qk += ((((tl.dot(q, k).to(tl.float32) * q_descale)) * k_descale) * Qk_scale)
         else:
             if INT8_KV:
@@ -163,9 +167,13 @@ def _attn_fwd_inner(
             # DO NOT CALL composed_dot_both.
             # The generated code will trigger https://github.com/ROCm/aotriton/issues/54
             # for BLOCK_M = 128 and BLOCK_N = 64
-            qk += (Qk_scale * tl.dot(q0, k0))
-            if BLOCK_DMODEL1 > 0 : qk += (Qk_scale * tl.dot(q1, k1))
-            if BLOCK_DMODEL2 > 0 : qk += (Qk_scale * tl.dot(q2, k2))
+            # out_dtype=tl.float32: explicit fp32 accumulation prevents Triton from
+            # keeping the dot result in the input dtype (bf16/fp16) before the
+            # Qk_scale multiply, which on gfx1151 can produce intermediate values
+            # that overflow or lose mantissa bits before being promoted.
+            qk += (Qk_scale * tl.dot(q0, k0, out_dtype=tl.float32))
+            if BLOCK_DMODEL1 > 0 : qk += (Qk_scale * tl.dot(q1, k1, out_dtype=tl.float32))
+            if BLOCK_DMODEL2 > 0 : qk += (Qk_scale * tl.dot(q2, k2, out_dtype=tl.float32))
 
         if B_ptrs is not None:
             NS = start_N + OFFS_N
